third_party/blink/web_tests/platform/generic/crypto/subtle/hmac/import-jwk-expected.txt

Test importing a JWK key for HMAC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a key...

PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 256
PASS key.usages is ["sign", "verify"]

Using the key to sign message 'foo'...
PASS bytesToHexString(new Uint8Array(signature)) is 'e03736fe098892b2a2da77812431f7c014d32e2fd69f3bcff883ac923a8fa2da'

Verifying the signature...
PASS verificationResult is true
PASS successfullyParsed is true

TEST COMPLETE
